requirement.txt
dotenv
flask
os
logging
dotenvcohere
flask_cors